+++ src/backend/plugins/video/warren_video/indicator/viewers/tests/conftest.py
@@ -1,0 +1,12 @@
+"""complete views fixtures."""
+
+# pylint: disable=unused-import
+
+from warren_video.tests.conftest import (  # noqa: F401
+    assert_matching_videos,
+    assert_no_matching_video,
+    non_mocked_hosts,
+)
+
+from warren.tests.fixtures.app import http_client  # noqa: F401
+from warren.tests.fixtures.asynchronous import anyio_backend  # noqa: F401