+++ src/backend/plugins/video/warren_video/indicator/views/__init__.py
@@ -1,0 +1,1 @@
+"""Views video indicator."""